feat : 주문 검증 로직
